# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: نظام إدارة توزيع السلع - Distribution Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: getStores
      tags: [stores]
      summary: Get all stores
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [stores]
      summary: Create store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreRequest"
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /stores/{id}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Get store by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Store details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
    put:
      operationId: updateStore
      tags: [stores]
      summary: Update store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoreRequest"
      responses:
        "200":
          description: Updated store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
    delete:
      operationId: deleteStore
      tags: [stores]
      summary: Delete store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get all tasks
      parameters:
        - name: distributorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, cancelled]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update task status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /suggestions:
    get:
      operationId: getStoreSuggestions
      tags: [suggestions]
      summary: Get store suggestions from distributors
      responses:
        "200":
          description: Store suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoreSuggestion"
    post:
      operationId: createStoreSuggestion
      tags: [suggestions]
      summary: Distributor suggests a new store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreSuggestionRequest"
      responses:
        "201":
          description: Suggestion submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSuggestion"
  /suggestions/{id}/approve:
    put:
      operationId: approveStoreSuggestion
      tags: [suggestions]
      summary: Admin approve store suggestion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suggestion approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSuggestion"
  /suggestions/{id}/reject:
    put:
      operationId: rejectStoreSuggestion
      tags: [suggestions]
      summary: Admin reject store suggestion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suggestion rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSuggestion"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Store:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
          nullable: true
        debt:
          type: number
        totalVisits:
          type: integer
        lastVisit:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - ownerName
        - phone
        - latitude
        - longitude
        - debt
        - totalVisits
        - createdAt
    CreateStoreRequest:
      type: object
      properties:
        name:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
      required:
        - name
        - ownerName
        - phone
        - latitude
        - longitude
    UpdateStoreRequest:
      type: object
      properties:
        name:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
    TaskItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - productId
        - productName
        - quantity
        - price
    Task:
      type: object
      properties:
        id:
          type: integer
        distributorId:
          type: integer
        distributorName:
          type: string
        storeId:
          type: integer
        storeName:
          type: string
        storeLatitude:
          type: number
        storeLongitude:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/TaskItem"
        status:
          type: string
          enum: [pending, in_progress, completed, failed, cancelled]
        totalAmount:
          type: number
        notes:
          type: string
          nullable: true
        failureReason:
          type: string
          nullable: true
        distance:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - distributorId
        - distributorName
        - storeId
        - storeName
        - storeLatitude
        - storeLongitude
        - items
        - status
        - totalAmount
        - createdAt
    CreateTaskRequest:
      type: object
      properties:
        distributorId:
          type: integer
        storeId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
            required:
              - productId
              - quantity
        notes:
          type: string
      required:
        - distributorId
        - storeId
        - items
    UpdateTaskRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed, failed, cancelled]
        notes:
          type: string
        failureReason:
          type: string
...
    StoreSuggestion:
      type: object
      properties:
        id:
          type: integer
        distributorId:
          type: integer
        distributorName:
          type: string
        name:
          type: string
        ownerName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        latitude:
          type: number
        longitude:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - distributorId
        - distributorName
        - name
        - latitude
        - longitude
        - status
        - createdAt
    CreateStoreSuggestionRequest:
      type: object
      properties:
        name:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        address:
          type: string
        photoUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
      required:
        - name
        - phone
        - address
        - latitude
        - longitude
...